Presentation of Mahesh Sabnis's Session
Session Tital - Need and Importance of Azure Key Vault
Speaker - MAHESH SABNIS
About session - Azure Key Vault is a cloud service for securely storing and accessing secrets. A secret is anything that you want to tightly control access to, such as API keys, passwords, certificates, or cryptographic keys. In this session we will discuss uses of it.